feat: centauri map
