+++ src/App.js
@@ -1,8 +1,9 @@
 import React, {useEffect, useState} from "react";
-import { BrowserRouter, Route, Routes } from "react-router-dom";
+import {BrowserRouter, Route, Routes} from "react-router-dom";
 import Home from "./Components/Home/Home.jsx";
 import About from "./Components/About/About.jsx";
 import Maps from "./Components/Maps/Maps.jsx";
+import CentauriMap from "./Components/Maps/Centauri/CentauriMap.jsx";
 import JoinUs from "./Components/Join-us/JoinUs.jsx";
 import Login from "./Components/Admin/Login/Login.jsx";
 import Background from "./Components/Layout/Background/Background.jsx";
@@ -19,9 +20,9 @@ const App = () => {
         }, 1500);
     }, []);
 
-    return(
+    return (
         <>
-            <Header />
+            <Header/>
             {loading ? (
                 <div className="loader__container">
                     <Loader/>
@@ -29,11 +30,12 @@ const App = () => {
             ) : (
                 <BrowserRouter>
                     <Routes>
-                        <Route path="/" element={ <Home/> } />
-                        <Route path="/about" element={ <About/> } />
-                        <Route path="/maps" element={ <Maps/> } />
-                        <Route path="/join-us" element={<JoinUs />} />
-                        <Route path="/login" element={<Login />} />
+                        <Route path="/" element={<Home/>}/>
+                        <Route path="/about" element={<About/>}/>
+                        <Route path="/maps" element={<Maps/>}/>
+                        <Route path="/centauri" element={<CentauriMap/>}/>
+                        <Route path="/join-us" element={<JoinUs/>}/>
+                        <Route path="/login" element={<Login/>}/>
                     </Routes>
                 </BrowserRouter>
             )}
