+++ src/Components/Cards/ModalCard/modalCard.scss
@@ -1,0 +1,126 @@
+/* The Modal (background) */
+.modal {
+  visibility: hidden;
+  transition: all 1s ease-out;
+  opacity: 0;
+  display: block;
+  position: fixed;
+  padding-top: 100px;
+  right: 100px;
+  top: 100px;
+  min-width: 500px;
+  width: 20%;
+  height: 100%;
+  overflow: auto;
+}
+
+/* Modal Content */
+.modal-content {
+  box-sizing: border-box;
+  width: 100%;
+  max-width: 800px;
+  height: auto;
+  background-color: #25292D;
+  border: 1px solid #70767B;
+  border-radius: 12px;
+  padding: 0 20px 40px 20px;
+  transition: all 0.5s ease-in-out;
+
+  &-title {
+    padding: 20px 0 20px 0;
+    font-family: Nasalization;
+    font-size: 25px;
+    color: #BBC3C9;
+    font-style: normal;
+    font-weight: 400;
+    text-align: center;
+  }
+
+  &-img {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+    border-radius: 12px;
+    margin-bottom: 35px;
+  }
+
+  &-firstText {
+    font-family: AllRoundGothic;
+    color: #BBC3C9;
+    font-style: normal;
+    font-weight: 200;
+    font-size: 18px;
+    line-height: 30px;
+    padding-bottom: 35px;
+  }
+
+  &-secondText {
+    font-family: AllRoundGothic;
+    color: #BBC3C9;
+    font-style: normal;
+    font-weight: 200;
+    font-size: 18px;
+    line-height: 40px;
+
+    &-list {
+      list-style-type: none;
+
+      &-item {
+        border-top: $light-gray solid 1px;
+        border-left: $light-gray solid 1px;
+        border-right: $light-gray solid 1px;
+        height: 45px;
+
+        &:first-child {
+          border-radius: 3px 3px 0 0;
+        }
+
+        &:last-child {
+          border-bottom: $light-gray solid 1px;
+          border-radius: 0 0 3px 3px;
+        }
+
+        &-gps {
+          width: 100%;
+          margin-left: 10px;
+
+          &-title {
+            width: 19%;
+          }
+
+          &-content {
+            width: 350px;
+            background: #22272d;
+            border: 1px solid #14171a;
+            border-radius: 3px;
+            border-left: 3px solid #396377;
+            color: $light-gray;
+            page-break-inside: avoid;
+            font-family: monospace;
+            font-size: 11px;
+            overflow: auto;
+            padding: 0.5em 1em 0.5em 1em;
+            margin-top: -5px;
+            margin-left: 15px;
+            word-wrap: break-word;
+            cursor: pointer;
+          }
+        }
+
+        &-text {
+          margin-left: 10px;
+
+          &-data {
+            background: #22272d;
+            border: 1px solid #14171a;
+            border-radius: 3px;
+            font-family: monospace;
+            font-size: 12px;
+            margin-left: 10px;
+            padding: 0.5em 1em;
+          }
+        }
+      }
+    }
+  }
+}
