+++ src/Components/Maps/Centauri/CentauriSVG.jsx
@@ -1,0 +1,94 @@
+import React from "react";
+
+const CentauriSVG = () => {
+    return (
+        <>
+            <g className="assets">
+                <circle cx="540" cy="538" r="82" stroke="url(#paint0_radial_135_48)"
+                        strokeWidth="22"
+                />
+                <circle cx="539.5" cy="538.5" r="94.5" stroke="url(#paint1_radial_135_48)"
+                        strokeWidth="22"
+                />
+                <circle cx="540" cy="538" r="125" stroke="url(#paint2_radial_135_48)"
+                        strokeWidth="22"
+                />
+                <circle cx="540" cy="538" r="156" stroke="url(#paint3_radial_135_48)"
+                        strokeWidth="30"
+                />
+                <circle cx="539.5" cy="538.5" r="177" stroke="url(#paint4_radial_135_48)"
+                        strokeWidth="37"
+                />
+                <circle cx="213" cy="937" r="30.5" stroke="#979797"/>
+                <circle cx="857" cy="202" r="32.5" stroke="#979797"/>
+                <circle cx="440" cy="139" r="29.5" stroke="#979797"/>
+                <circle cx="337" cy="240" r="32.5" stroke="#979797"/>
+
+                <circle cx="539.5" cy="538.5" r="202" stroke="white" strokeOpacity="0.5"/>
+                <circle cx="539.5" cy="538.5" r="245" stroke="white" strokeOpacity="0.5"/>
+                <circle cx="540" cy="540" r="283.5" stroke="white" strokeOpacity="0.5"/>
+                <circle cx="540" cy="540" r="323.5" stroke="white" strokeOpacity="0.5"/>
+                <circle cx="538" cy="538" r="361.5" stroke="white" strokeOpacity="0.5"/>
+                <circle cx="537.5" cy="540.5" r="412" stroke="white" strokeOpacity="0.5"/>
+                <circle cx="540" cy="540" r="465.5" stroke="white" strokeOpacity="0.5"/>
+                <circle cx="539.5" cy="538.5" r="515" stroke="white" strokeOpacity="0.5"/>
+            </g>
+            <circle cx="213" cy="937" r="24" stroke="url(#paint5_radial_135_48)"
+                    strokeWidth="10"
+            />
+            <defs>
+                <radialGradient id="paint0_radial_135_48" cx="0" cy="0" r="1"
+                                gradientUnits="userSpaceOnUse"
+                                gradientTransform="translate(540 538) rotate(90) scale(232.525)"
+                >
+                    <stop/>
+                    <stop offset="1" stopColor="#F09000" stopOpacity="0"/>
+                </radialGradient>
+                <radialGradient id="paint1_radial_135_48" cx="0" cy="0" r="1"
+                                gradientUnits="userSpaceOnUse"
+                                gradientTransform="translate(539.5 538.5) rotate(90) scale(273.463)"
+                >
+                    <stop/>
+                    <stop offset="1" stopColor="#F09000" stopOpacity="0"/>
+                </radialGradient>
+                <radialGradient id="paint2_radial_135_48" cx="0" cy="0" r="1"
+                                gradientUnits="userSpaceOnUse"
+                                gradientTransform="translate(540 538) rotate(90) scale(373.35)"
+                >
+                    <stop/>
+                    <stop offset="1" stopColor="#F09000" stopOpacity="0"/>
+                </radialGradient>
+                <radialGradient id="paint3_radial_135_48" cx="0" cy="0" r="1"
+                                gradientUnits="userSpaceOnUse"
+                                gradientTransform="translate(540 538) rotate(90) scale(461.775)"
+                >
+                    <stop/>
+                    <stop offset="1" stopColor="#F09000" stopOpacity="0"/>
+                </radialGradient>
+                <radialGradient id="paint4_radial_135_48" cx="0" cy="0" r="1"
+                                gradientUnits="userSpaceOnUse"
+                                gradientTransform="translate(539.5 538.5) rotate(90) scale(519.088)"
+                >
+                    <stop/>
+                    <stop offset="1" stopColor="#F09000" stopOpacity="0"/>
+                </radialGradient>
+                <radialGradient id="paint5_radial_135_48" cx="0" cy="0" r="1"
+                                gradientUnits="userSpaceOnUse"
+                                gradientTransform="translate(213 937) rotate(90) scale(32.0625)"
+                >
+                    <stop/>
+                    <stop offset="1" stopColor="#BBBBBB" stopOpacity="0"/>
+                </radialGradient>
+                <radialGradient id="paint6_radial_135_48" cx="0" cy="0" r="1"
+                                gradientUnits="userSpaceOnUse"
+                                gradientTransform="translate(540 538) rotate(90) scale(197)"
+                >
+                    <stop/>
+                    <stop offset="1" stopColor="#876533" stopOpacity="0"/>
+                </radialGradient>
+            </defs>
+        </>
+    )
+}
+
+export default CentauriSVG;
