+++ src/Components/Maps/Centauri/centauriMap.scss
@@ -1,0 +1,98 @@
+@mixin planetHover {
+  transition: all 0.2s ease-in-out;
+  &:hover {
+    filter: brightness(150%);
+  }
+}
+
+/* The Maps */
+.maps {
+  display: flex;
+
+  &-svg {
+    &-avernus {
+      &-planet {
+        @include planetHover;
+      }
+    }
+
+    &-triton {
+      &-planet {
+        @include planetHover;
+      }
+    }
+
+    &-overvent {
+      &-planet {
+        @include planetHover;
+      }
+    }
+
+    &-rothis {
+      &-planet {
+        @include planetHover;
+      }
+    }
+
+    &-terra {
+      &-planet {
+        @include planetHover;
+      }
+    }
+
+    &-exolia {
+      &-planet {
+        @include planetHover;
+      }
+    }
+
+    &-treossa {
+      &-planet {
+        @include planetHover;
+      }
+    }
+
+    &-pertam {
+      &-planet {
+        @include planetHover;
+      }
+    }
+
+    &-titan {
+      &-planet {
+        @include planetHover;
+      }
+    }
+
+    &-luna {
+      &-planet {
+        @include planetHover;
+      }
+    }
+
+    &-mars {
+      &-planet {
+        @include planetHover;
+      }
+    }
+
+    &-europa {
+      &-planet {
+        @include planetHover;
+      }
+    }
+
+    &-alien {
+      &-planet {
+        @include planetHover;
+      }
+    }
+
+    &-wormhole {
+      &-planet {
+        @include planetHover;
+      }
+    }
+  }
+}
+
