--- src/translation/map.json
@@ -1,209 +1,0 @@
-{
-
-
-    "copy":"Copié!",
-
-    "titles":{
-        "Centauri":"Système Centauri",
-        "gps-title":"GPS :",
-        "size-title":"Taille :",
-        "gravity-title":"Gravité :",
-        "atmosphere-title":"Atmosphère/Respirable :",
-        "ore-title":"Minerais :",
-        "ice-title":"Glace :",
-        "fauna-title":"Faune :",
-        "weather-title":"Météo :",
-        "temperature-title":"Température :"
-    },
-
-    "planets":{
-        "avernus":{
-            "title":"Avernus",
-            "image":"planets/avernus.png",
-            "description":"",
-            "gps":"GPS:• Avernus:32025.97:322571.41:5673827.54:#FF776239:",
-            "size":"120 km",
-            "gravity":"1 g",
-            "atmosphere":"1 atm/oui",
-            "ore":"( Fe, Ni, Si, Co, Mg, Ar, Au )",
-            "ice":"Oui",
-            "fauna":"?",
-            "weather":"?",
-            "temperature":"?"
-        },
-        "triton":{
-            "title":"Triton",
-            "image":"planets/triton.png",
-            "description":"",
-            "gps":"GPS:• Triton:-6879522.35:-240237.73:-3220938.96:#FFC4F0F1:",
-            "size":"80.25 km",
-            "gravity":"1 g",
-            "atmosphere":"Oui/Oui",
-            "ore":"( Si, Fe, Ni, Mg, Au, Ag, Co )",
-            "ice":"Oui",
-            "fauna":"Aucune",
-            "weather":"Blizzard",
-            "temperature":"Glaciale"
-        },
-        "overvent":{
-            "title":"Overvent",
-            "image":"planets/overvent.png",
-            "description":"",
-            "gps":"GPS:• Overvent:0.5:0.5:0.5:#FF9D442E:",
-            "size":"1000 km",
-            "gravity":"6.4 g",
-            "atmosphere":"Non",
-            "ore":"Aucun",
-            "ice":"Aucune",
-            "fauna":"Aucune",
-            "weather":"Enfer",
-            "temperature":"?"
-        },
-        "rothis":{
-            "title":"Rothis",
-            "image":"planets/rothis.png",
-            "description":"",
-            "gps":"GPS:• Rothis:-9169841.25:-251041.15:-1258370.22:#FF75A1F1:",
-            "size":"22 km",
-            "gravity":"0.27 g",
-            "atmosphere":"0.2 atm / non",
-            "ore":"( Fe, Co, Ni, Mg, Si, Ag, Au, Pt )",
-            "ice":"Oui",
-            "fauna":"Aucun",
-            "weather":"Non",
-            "temperature":"Glaciale"
-        },
-        "terra":{
-            "title":"Terra",
-            "image":"planets/terra.png",
-            "description":"",
-            "gps":"GPS:• Terra:-945.97:-601342.12:-9681342:#FF75F1AB:",
-            "size":"120 km",
-            "gravity":"1 g",
-            "atmosphere":"1 atm/oui",
-            "ore":" ( Si, Mg, Fe, Co, Ni, Au, Ag )",
-            "ice":"Oui",
-            "fauna":"Loups",
-            "weather":"Variée",
-            "temperature":"Tempérée"
-        },
-        "exolia":{
-            "title":"Exolia",
-            "image":"planets/exolia.png",
-            "description":"",
-            "gps":"GPS:• Exolia:-9136993.11:-168166.36:-1061633.96:#FF95FF7C:",
-            "size":"112.8 km",
-            "gravity":"0.98 g",
-            "atmosphere":"0.8 atm / oui",
-            "ore":"( Fe, Ni, Si, Co, Ag, Mg, Au )",
-            "ice":"Oui",
-            "fauna":"Oui",
-            "weather":"variable (pas d'orage)",
-            "temperature":"Tropicale (froid la nuit, tempéré le jour)"
-        },
-        "treossa":{
-            "title":"Treossa ",
-            "image":"planets/treossa.png",
-            "description":"Ils l'appellent la planète Moisson, la seule chose que cette planète n'a pas c'est la glace et l'oxygène, elle a une atmosphère mais n'est pas respirable... Quand vous allez visiter cette planète, n'oubliez pas de prendre assez d'oxygène et d'hydrogène...",
-            "gps":"GPS:• Treossa:4142755.46:265722.33:4605103.46:#FFF17575:",
-            "size":"120 km",
-            "gravity":"1 g",
-            "atmosphere":"1 atm / non",
-            "ore":"( Fe, Ni, Si, Co, Ar, Mg, U, Au, Pt )",
-            "ice":"Non",
-            "fauna":"?",
-            "weather":"?",
-            "temperature":"?"
-        },
-        "pertam":{
-            "title":"Pertam",
-            "image":"planets/pertam.png",
-            "description":"",
-            "gps":"GPS:• Pertam:-6198476.78:-195591.28:-5812969.76:#FFF1C875:",
-            "size":"60.13 km",
-            "gravity":"1.20 g",
-            "atmosphere":"oui / oui",
-            "ore":"( Fe, Ni, Si, Co, Mg, Ag, Au )",
-            "ice":"Oui",
-            "fauna":"Aucune",
-            "weather":"Tempêtes de sable, Orages",
-            "temperature":"Chaude"
-        },
-        "titan":{
-            "title":"Titan",
-            "image":"planets/titan.png",
-            "description":"",
-            "gps":"GPS:• Titan:8000000.5:0.5:-169999.5:#FF75C9F1:",
-            "size":"19 km",
-            "gravity":"0.25 g",
-            "atmosphere":"Oui / Non",
-            "ore":"( Ni, Fe, Si, Mg, Pt, Au, Co, Ag )",
-            "ice":"Oui",
-            "fauna":"Aucune",
-            "weather":"?",
-            "temperature":"?"
-        },
-        "luna":{
-            "title":"Luna",
-            "image":"planets/luna.png",
-            "description":"",
-            "gps":"GPS:• Luna:4847.01:-607289.14:-9881066.74:#FFA1A1A1:",
-            "size":"19 km",
-            "gravity":"0.25 g",
-            "atmosphere":"non",
-            "ore":"( Ni, Fe, Si, Mg, Pt, Au, Co, Ag )",
-            "ice":"Oui",
-            "fauna":"Aucune",
-            "weather":"Aucune",
-            "temperature":"Glaciale"
-        },
-        "mars":{
-            "title":"Mars",
-            "image":"planets/mars.png",
-            "description":"",
-            "gps":"GPS:• Mars:3025820.43:-466517.93:-8356839.47:#FFCC756A:",
-            "size":"120 km",
-            "gravity":"0.90 g",
-            "atmosphere":"Oui/Non",
-            "ore":"( Si, Mg, Fe, Ni, Au, Ag, Co )",
-            "ice":"Oui",
-            "fauna":"Aucune",
-            "weather":"?",
-            "temperature":"?"
-        },
-        "europa":{
-            "title":"Europa",
-            "image":"planets/europa.png",
-            "description":"",
-            "gps":"GPS:• Europa:3123694.62:-487168.58:-8556864.11:#FF75F1DF:",
-            "size":"19 km",
-            "gravity":"0.25 g",
-            "atmosphere":"Oui/Non",
-            "ore":"( Ni, Fe, Si, Mg, Pt, Au, Co, Ag )",
-            "ice":"Oui",
-            "fauna":"Aucune",
-            "weather":"?",
-            "temperature":"?"
-        },
-        "alien":{
-            "title":"Alien",
-            "image":"planets/alien.png",
-            "description":"",
-            "gps":"GPS:• Alien:8000000.5:0.5:0.5:#FFE2A8E2:",
-            "size":"120 km",
-            "gravity":"1.10 g",
-            "atmosphere":"Oui/Non",
-            "ore":"( Si, Mg, Fe, Ni, Au, Ag, Co )",
-            "ice":"Oui",
-            "fauna":"Aucune",
-            "weather":"?",
-            "temperature":"?"
-        },
-        "wormhole":{
-            "title":"Wormhole",
-            "image":"planets/wormhole.png",
-            "description":"",
-            "gps":"GPS:Athena Orion:-2958050:-417028:-3668261:#FF176AAD:"
-        }
-    }
-}
